+++ node/CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.  The format
 
 ## [Unreleased]
 
+### Added
+* Introducing fast-syncing to join the network, avoiding the need to execute every block to catch up.
+* Added `archival_sync` to `[node]` config section, along with archival syncing capabilities
+
 ### Changed
 * Support building and testing using stable Rust.
 * Log chattiness in `debug` or lower levels has been reduced and performance at `info` or higher slightly improved.